feat(withFilterField): new Hook for easy filtering
