+++ package.json
@@ -39,6 +39,7 @@
     "@nestjs/core": "^9.0.0",
     "@nestjs/platform-express": "^9.0.0",
     "content-range": "^2.0.2",
+    "flat": "^5.0.2",
     "lodash": "^4.17.21",
     "luxon": "^3.1.1",
     "reflect-metadata": "^0.1.13",
@@ -57,6 +58,7 @@
     "@semantic-release/npm": "^9.0.1",
     "@semantic-release/release-notes-generator": "^10.0.3",
     "@types/express": "^4.17.13",
+    "@types/flat": "^5.0.2",
     "@types/jest": "28.1.8",
     "@types/node": "^16.0.0",
     "@types/supertest": "^2.0.11",
